Shared equatable PaymentMethods
