+++ Scripts/test-carthage-integration.sh
@@ -134,6 +134,7 @@ cp "../Tests/Card Tests/3DS2 Component/ThreeDSResultExtension.swift" Tests/Three
 cp "../Tests/Helpers/XCTestCase+RootViewController.swift" Tests/XCTestCase+RootViewController.swift
 cp "../Tests/Helpers/XCTestCase+Wait.swift" Tests/XCTestCase+Wait.swift
 cp "../Tests/Helpers/UIViewController+Search.swift" Tests/UIViewController+Search.swift
+cp "../Tests/Helpers/PaymentMethods+Equatable.swift" Tests/PaymentMethods+Equatable.swift
 cp "../Tests/Adyen Tests/Analytics/AnalyticsProviderMock.swift" Tests/AnalyticsProviderMock.swift
 cp "../Tests/DummyData/Dummy.swift" Tests/Dummy.swift
 cp "../Demo/Common/Utils/APIClientMock.swift" Tests/APIClientMock.swift
